Removed line interrupt and HPEN polling
The hope was that the line interrupt could give us a second chance at a missed frame interrupt, but it never really helped.  It usually meant we were late for the next frame too.

When there was little to draw we could catch the line interrupt in the same frame, so we needed to check HPEN to make sure we were beyond it.  This nasty hack has been removed too.