Improved tile updating logic
Changed to update all tiles if no sprites are visible, which makes the maze appear faster and reduces artefacts on the attract screen text.  When sprites are visible we fall back on the original strip method.

As a bonus, it's now much easier to change the code for the Mayhem accelerator, and it may even be possible to detect and patch the code automatically.