Improved control support for diagonal inputs
The original Pac-Man machine uses a 4-way joystick, which makes it
impossible to generate more than one direction input. The game does
not understand diagonals, and will use the first direction found to
be pressed. This can cause frustration when played on the keyboard
or with a regular 8-way joystick.

We now ignore control input for the direction Pac-Man is currently
facing, so changes of direction are always preferred. We also ignore
perpendicular inputs on tile entry, to defeat the turn overshoot
mechanism that gives a brief window for players to change their mind
about a turn.

These changes are only active when diagonal inputs are detected, so
traditional play with a single input remains true to the original.